+++ src/setting/Size.java
@@ -1,0 +1,9 @@
+package setting;
+
+public class Size {
+
+    public static int boardWidth = 500;
+    public static int boardHeight = 900;
+    public static int fontSize = 23;
+
+}
